All conflicts resolved in favor of 0.99. (Conflicts had arisen from bugfixes previously cherry-picked for v0.98b/c.)

Active development will continue on branch 1.00, with bugfixes for 0.99b (if necessary) cherry-picked on the master branch.

+++ CvGameCoreDLL/CyCity.h
@@ -182,7 +182,7 @@ public:
 	int getPopulation();
 	void setPopulation(int iNewValue);
 	void changePopulation(int iChange);
-	long getRealPopulation();
+	int getRealPopulation();
 
 	int getHighestPopulation();
 	void setHighestPopulation(int iNewValue);
